modification: moving ui records
moving records specific ui files  into subfolder records
